requirements.txt
alive_progress
duckdb
python-levenshtein
